More changes. Added sensors, and larger kidneys. Fixed client-side food bar.

+++ com/immibis/fluffyjam1/Reagents.java
@@ -160,7 +160,7 @@ public class Reagents implements Serializable {
 		return capacity - total;
 	}
 
-	public double getFractionFull() {
+	public float getFractionFull() {
 		return total / capacity;
 	}
 }
